Refactor some garbage

+++ src/test/java/edu/pdx/cs410J/bspriggs/TextDumperTest.java
@@ -34,7 +34,7 @@ public class TextDumperTest {
     protected static Random r = new Random();
 
     public static Date generateDate() {
-        return new Date(r.nextLong());
+        return new Date(r.nextInt());
     }
 
     public static String generatePhoneNumber() {
